Add `AppRuleExt::replicate_bundle_with`
Allows customizing the priority of a bundle. This requires manually
specifying the priority, but since changing the number of replicated
components in a bundle would require updating the trait anyway, it's not
a big drawback. I also expect manual impl to be rarely needed, at least
less often than setting the priority during registration.