Migrate to Bevy states
This allows users to utilize things like `StateScoped` and run systems
more efficiently using `OnEnter` or `OnExit` without evaluating
conditions every frame.

As a result, we now require `StatesPlugin` to be present. It's included
by default in `DefaultPlugins`, but with `MinimalPlugins` you have to
add it manually. In tests, I had to add `StatesPlugin` everywhere.