Fix replication of empty entities before connection (#584)
Entities with components aren't affected because we check the last
mutation tick. But for empty entities, we only check `marker_added`,
which is true only for entities spawned during this tick. I added
`empty_before_connection` which tirggers this.

Checking the mutation tick fixes the problem. Moreover, this check is
sufficient - we don't need to check the marker. I inspected our tests
and added 2 more to ensure it's actually the case.