Component-based visibility (#595)
With observers, we can finally provide an ergonomic visibility API. This
works similarly to collision layers in physics: users insert filters
into both the client and gameplay entities.

An entity will be visible to a client if it has all the filter
components present on the entity, and `VisibilityFilter::is_visible`
returns `true` for each of them.

I used a separate trait instead of `PartialEq` for flexibility. For
example, with a custom trait, users can use bitmasks. I also considered
providing a blanket impl for `PartialEq`, but this would prevent users
from having both `PartialEq` and a custom implementation for bitflags.

I am planning to use this system to add per-component visibility later.

---------

Co-authored-by: UkoeHB <37489173+UkoeHB@users.noreply.github.com>