Track sent components for each entity
We'll need this for per-component visibility, but this change is useful
on its own because it fixes replication for rules with multiple
components that were inserted on different ticks.

For example, you have a replication rule for `(A, B)`. You spawn an
entity with `A`, and on the next tick you insert `B`. Without this
change, `B` is replicated, but `A` is wrongly considered as sent. I
added a test for it.