Track sent components for each entity (#604)
* Track sent components for each entity

We'll need this for per-component visibility, but this change is useful
on its own because it fixes replication for rules with multiple
components that were inserted on different ticks.

For example, you have a replication rule for `(A, B)`. You spawn an
entity with `A`, and on the next tick you insert `B`. Without this
change, `B` is replicated, but `A` is wrongly considered as sent. I
added a test for it.

* Add `SerdeFns` wrapper to simplify code

Reduces the unsafety and the mount of passed arguments.

* Replace hash lookups with a dynamically growing bitmask.

This significantly increases performances.

---------

Co-authored-by: UkoeHB <37489173+UkoeHB@users.noreply.github.com>