Parallelize replication sending
Our replication send system is huge. To make it more readable, it was
split into manageable helpers.

But the system contains 15 system parameters, which is the maximum in
Bevy. If I want to add more, I need to group parameters into tuples,
which isn't nice.

Since it was already split into helpers internally, I decided to turn
those helpers into systems and chain them. I originally considered
piping them, but it doesn't work well with returning a result. Also,
according to my benchmarks, it's a bit faster when there are many
clients. This also reduces the number of components in client queries.

I had to turn `SerializedData` into a resource and move its cleanup into
`send_messages`. I also created a resource that holds the change tick,
since it needs to be shared between two systems.