Do not resend messages locally after disconnect
In the RTS example, when the server disconnects a client because it
selected an already occupied team, protocol events were being resent
locally on the client.

This happened because events are resent locally when the client is not
connected, to support single-player and listen-server modes.

To fix this, we need toreset events after disconnect. Previously, events
were reset before connection, but doing so after disconnect is more
consistent with the rest of our logic.

Added two tests reproducing the issue (messages and events) and verified
they now pass.