Add per-component visibility (#614)
* Refactor `ComponentMask`

Make the API similar to `bitflags`. This also makes removals a bit more
efficient.

* Add per-component visibility

The `VisibilityFilter` trait now includes an associated `Scope` type,
which defines whether the filter affects the entire entity or specific
components. It can be `Entity` to hide the whole entity, a
`ComponentScope` to hide a single component, or a tuple of components
to hide multiple components. `ComponentScope` is needed only because of
the orphan rule: we can't have a blanket impl for any `Component` and
for tuples of components.

The data for each filter is stored in the `FilterRegistry` with quick
access via the filter bit. Component sets for filters are stored as a
`ComponentMask` to avoid hash lookups for each component. I also had to
make it public since it's used in the return value of the `FilterScope`
trait.

Since filter bits are now used outside the `visibility` module, I
created dedicated types for them with convenient methods.

---------

Co-authored-by: UkoeHB <37489173+UkoeHB@users.noreply.github.com>